mini-project/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
transformers>=4.44.0
torch>=2.3.0
pydantic>=2.8.0
python-dotenv>=1.0.1
requests>=2.32.3
jinja2>=3.1.4
reportlab>=4.2.2
tqdm>=4.66.0
beautifulsoup4>=4.12.3
tavily-python>=0.5.0
pypdf>=6.9.2
playwright==1.58.0
markdown==3.10.2